chore(애니메이션):'재단 입장 연출"
