--- Assets/02. Scripts/BossController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 84bbe2c931bdd1f46a88ab45fcf310ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
