--- Assets/02. Scripts/MonsterController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 316f22322c27c2144aa6468b44ea9ac1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
