+++ Assets/02. Scripts/UI/VignetteIntensityAnimator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 558bcaec192f7cb4e86a72f4c94cb1ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
