+++ Assets/05. Shaders/Volume/V_Lanterns.asset
@@ -33,7 +33,7 @@ MonoBehaviour:
     m_Value: 11
   aperture:
     m_OverrideState: 1
-    m_Value: 1
+    m_Value: 28.1
   focalLength:
     m_OverrideState: 0
     m_Value: 1
