+++ Assets/07. Prefabs/LevelPref/Altar.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dbc4ac68a1a67814f8f62e4e5e79979e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
