+++ Assets/07. Prefabs/LevelPref/BossBat (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3d43ed24f7a83234389df7c4e4d792c6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
