+++ Assets/07. Prefabs/LevelPref/BossBat.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 259fff81db6149948b8d68eb4b872f27
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
