+++ Assets/07. Prefabs/LevelPref/DropModel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2f1d96bbe1950ee438560c3e55b0cd17
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
