+++ Assets/07. Prefabs/LevelPref/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0c767a9dbe00764f9a409386d627ae1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
