+++ Assets/07. Prefabs/LevelPref/Slate.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 699ff1074bf96c14e81a416f68ee39e7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
