+++ Assets/07. Prefabs/LevelPref/Slate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 68d77c81fadea0c4cb288f0ea4f7ac4b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
