+++ Assets/07. Prefabs/LevelPref/Up.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 83555c47c695bdb43903df59a1a0a1a3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
