+++ Assets/09. Animations/Animation/AnimSD/IdleB.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 23cd5d5efff30f4439486e32bb29cf4a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
