+++ Assets/09. Animations/Animator/GrabSlate.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be97aaead2dd897408899725d4e39c30
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
