+++ Assets/13.Timeline/CutScene/Ingame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e60006b8d4e8024397bd35cc78ffddc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
