+++ Assets/13.Timeline/CutScene/Ingame/AltarEnter.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9ee6409de737e84d89a9ece4e232464
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
