+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,25 +5,25 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
+  - enabled: 1
     path: Assets/01. Scenes/Intro.unity
     guid: 5d7123ca8685d9c448b8e3ddfe2152ba
-  - enabled: 0
+  - enabled: 1
     path: Assets/01. Scenes/Title.unity
     guid: 5b9dfb77b2f9a404581101ca11b28b95
-  - enabled: 0
+  - enabled: 1
     path: Assets/01. Scenes/CutScene/GameStart.unity
     guid: c17028004e2d5204a983dcdfbfb3a7bf
-  - enabled: 0
+  - enabled: 1
     path: "Assets/01. Scenes/\uD50C\uB77C\uC6CC - 1.unity"
     guid: 6fcdea9cb45dd6045afaf841d1d9d15f
-  - enabled: 0
+  - enabled: 1
     path: "Assets/01. Scenes/\uD50C\uB77C\uC6CC - 2.unity"
     guid: 68a4588b025393d4791642821bfd56ae
-  - enabled: 0
+  - enabled: 1
     path: "Assets/01. Scenes/\uD50C\uB77C\uC6CC - 3.unity"
     guid: d38b757d28e412a4d954c12a812b5919
-  - enabled: 1
+  - enabled: 0
     path: Assets/11. Debug/S_Flower Test.unity
     guid: 38245888337dfb342a5172bc954e408d
   m_configObjects: {}
